supports checkpoints only feature (single-checkpoint)
