+++ hdfscontents/hdfsio.py
@@ -16,6 +16,7 @@ from notebook.utils import (
 import nbformat
 
 from pydoop.hdfs.path import split
+from pydoop.hdfs.fs import hdfs as HDFS
 from ipython_genutils.py3compat import str_to_unicode
 from traitlets.config import Configurable
 from traitlets import Bool, Integer, Unicode, default, Instance
@@ -41,14 +42,24 @@ def path_to_invalid(path):
 def hdfs_copy_file(hdfs, src, dst):
     chunk = 2 ** 16
     # TODO: check if we need to specify replication
-    with hdfs.open_file(dst, 'w') as f1:
-        with hdfs.open_file(src, 'r') as f2:
-            while True:
-                out = f2.read(chunk)
-                if len(out) == 0:
-                    break
-                f1.write(out)
-    hdfs.chmod(dst, 0o0770)
+    try:
+        with hdfs.open_file(dst, 'w') as f1:
+            with hdfs.open_file(src, 'r') as f2:
+                while True:
+                    out = f2.read(chunk)
+                    if len(out) == 0:
+                        break
+                    f1.write(out)
+    except:
+        with hdfs.open_file(dst, 'w') as f1:
+            with open(src, 'rb') as f2:
+                while True:
+                    out = f2.read(chunk)
+                    if len(out) == 0:
+                        break
+                    f1.write(out)
+
+    hdfs.chmod(dst, 0o0777)
 
 
 def hdfs_replace_file(hdfs, src, dst):
@@ -57,7 +68,7 @@ def hdfs_replace_file(hdfs, src, dst):
     """
     hdfs.delete(dst)
     hdfs.move(src, hdfs, dst)
-    hdfs.chmod(dst, 0o0770)
+    hdfs.chmod(dst, 0o0777)
 
 
 def hdfs_file_exists(hdfs, hdfs_path):
@@ -95,8 +106,8 @@ def atomic_writing(hdfs, hdfs_path):
     # Flush to disk
     fileobj.flush()
     fileobj.close()
-    hdfs.chmod(hdfs_path, 0o0770)
-    
+    hdfs.chmod(hdfs_path, 0o0777)
+
     # Written successfully, now remove the backup copy
     if hdfs_file_exists(hdfs, tmp_path):
         hdfs.delete(tmp_path)
@@ -130,7 +141,7 @@ def _simple_writing(hdfs, hdfs_path):
     # Flush to disk
     fileobj.flush()
     fileobj.close()
-    hdfs.chmod(hdfs_path, 0o0770)
+    hdfs.chmod(hdfs_path, 0o0777)
 
 
 class HDFSManagerMixin(Configurable):
@@ -149,10 +160,23 @@ class HDFSManagerMixin(Configurable):
     log : logging.Logger
     """
 
+    hdfs_namenode_host = Unicode(u'localhost', config=True, help='The HDFS namenode host')
+    hdfs_namenode_port = Integer(9000, config=True, help='The HDFS namenode port')
+    hdfs_user = Unicode(None, allow_none=True, config=True, help='The HDFS user name')
+
+    root_dir = Unicode(u'/', config=True, help='The HDFS root directory to use')
+
+    # The pydoop HDFS connection object used to interact with HDFS cluster.
+    hdfs = Instance(HDFS, config=True)
+
     use_atomic_writing = Bool(True, config=True, help=
     """By default notebooks are saved on disk on a temporary file and then if succefully written, it replaces the old ones.
-      This procedure, namely 'atomic_writing', causes some bugs on file system whitout operation order enforcement (like some networked fs).
-      If set to False, the new notebook is written directly on the old one which could fail (eg: full filesystem or quota )""")
+    This procedure, namely 'atomic_writing', causes some bugs on file system whitout operation order enforcement (like some networked fs).
+    If set to False, the new notebook is written directly on the old one which could fail (eg: full filesystem or quota )""")
+
+    @default('hdfs')
+    def _default_hdfs(self):
+        return HDFS(host=self.hdfs_namenode_host, port=self.hdfs_namenode_port, user=self.hdfs_user)  # groups=None
 
     def _hdfs_dir_exists(self, hdfs_path):
         """Does the directory exists in HDFS filesystem?
@@ -181,7 +205,7 @@ class HDFSManagerMixin(Configurable):
         if not self.hdfs.exists(hdfs_path):
             try:
                 self.hdfs.create_directory(hdfs_path)
-                self.hdfs.chmod(hdfs_path, 0o0770)
+                self.hdfs.chmod(hdfs_path, 0o0777)
             except OSError as e:
                 if e.errno != errno.EEXIST:
                     raise
@@ -236,7 +260,7 @@ class HDFSManagerMixin(Configurable):
         self.hdfs.move(src, self.hdfs,  dst)
         # The pydoop move changes the permissions back to default!
         for p in self.hdfs.walk(dst):
-            self.hdfs.chmod(p['name'], 0o0770)
+            self.hdfs.chmod(p['name'], 0o0777)
 
     def _hdfs_copy_file(self, src, dst):
         hdfs_copy_file(self.hdfs, src, dst)
